requirements.txt
python-dotenv
python-telegram-bot==13.14
requests
urllib3
